removed a utilities package and pushed the functions into the machine as required and changed the links to them.
